+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Test and beta releases will have experimental functions enabled by default.
 **Next release**
 
 * Show restrictions of system applications, even when disabled
+* Moved database back due to permission problems ([issue](https://github.com/M66B/XPrivacy/issues/1129))
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)